ensure single logical subscription per subject

+++ jestGlobalSetup.js
@@ -5,10 +5,16 @@ if (process.env.LEAKED_HANDLES) {
 }
 const {execSync} = require('child_process');
 
+let listenerAdded = false;
+
 module.exports = function () {
-  process.on('SIGINT', () => {
-    execSync(`bash -c "docker ps | grep jest-test-container | awk '{print \\$1}' | xargs -I{} docker stop {}"`);
-    execSync(`bash -c "docker ps -a | grep jest-test-container | awk '{print \\$1}' | xargs -I{} docker rm {}"`);
-    process.exit(0);
-  });
+  if (!listenerAdded) {
+    listenerAdded = true;
+    console.log('Configuring exit listener');
+    process.on('exit', () => {
+      console.log('Removing test containers');
+      execSync(`bash -c "docker ps | grep jest-test-container | awk '{print \\$1}' | xargs -I{} docker stop {}"`);
+      execSync(`bash -c "docker ps -a | grep jest-test-container | awk '{print \\$1}' | xargs -I{} docker rm {}"`);
+    });
+  }
 };
